+++ scripts/build_site.sh
@@ -8,7 +8,7 @@ echo TODO > site/generated/themes.md
 
 # Build the API docs, only build JSON output here to remove ~2s from
 # the time it takes to run this script.
-node bin/typedoc --json docs/docs.json
+node bin/typedoc --json docs/docs.json --readme none
 
 # Build the actual site, references the API docs
 node bin/typedoc --options site/typedoc.config.jsonc