+++ src/lib/internationalization/index.ts
@@ -1,9 +1,3 @@
-/**
- * Internationalization module, see {@link "Adding Locales" | Adding Locales} for details on extending TypeDoc.
- *
- * @module
- * @document ../../../internal-docs/internationalization.md
- */
 export {
     type TranslatableStrings,
     Internationalization,